improve logic, refactor and test routes
