--- src/main.py
@@ -1,2 +1,0 @@
-from src.api.DataFetcher import DataFetcher
-
